// packages/grab/src/types.ts
export interface DiscardPromptProps {
	label?: string;
	showCancel?: boolean;
	cancelOnEscape?: boolean;
	onConfirm?: () => void;
	onCancel?: () => void;
	onCopy?: () => void;
}